# swaggerui/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Library API - OpenAPI 3.0
  description: List and borrow book from library
  termsOfService: http://swagger.io/terms/
  contact:
    email: user@email.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080
  - url: https://library-api-prod.com:8080
...
security:
  - BearerAuth: []
paths:
...
  /borrowing-records/{id}:
    put:
      tags:
        - borrowing-record
      summary: Post book to return
      description: Post book to return
      operationId: returnBook
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnBorrowingRecordRes'

        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

        '401':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
components:
  schemas:
...
    ReturnBorrowingRecordRes:
      type: object
      properties:
        data:
          type: object
          example:
            id: 1
            userID: 1,
            bookID: 5,
            status: RETURNED
            borrowingDate: 2022-08-05T18:59:34.004384Z
            returningDate": 2022-08-08T20:35:34.863325635+07:00
...
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        code:
          type: string
          example: "UNAUTHORIZED_ERROR"
        message:
          type: string
          example: Unauthorized error
    InternalServerError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 500
        code:
          type: string
          example: "INTERNAL_SERVER_ERROR"
        message:
          type: string
          example: Internal server error
    BadRequestError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        code:
          type: string
          example: "BAD_REQUEST_ERROR"
        message:
          type: string
          example: Bad request error
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer